Split the integration test
